Add option to disable checking if link_name exists
The host lavapipe/swiftshader vulkan libraries and ICD manifest
json files historically exist in a nested aarch64-linux-gnu or
x86_64-linux-gnu subdirectory depending on the host arch.

Instead of adding more complex logic to have a link_name per
host arch, let's just add to option to skip the check and then
debian_substituion_marker can be updated with:

```
symlinks: {
  target: "/usr/lib/cuttlefish-common/bin/libvk_swiftshader.so"
  link_name: "bin/aarch64-linux-gnu/vulkan.pastel.so"
}
symlinks: {
  target: "/usr/lib/cuttlefish-common/bin/libvk_swiftshader.so"
  link_name: "bin/x86_64-linux-gnu/vulkan.pastel.so"
}
```

to create 2 links (one of which would never be used).

After the host migration is fully rolled out and stable,
crosvm_manager.cpp can be to reference the new vulkan libraries
and vulkan manfiest files directly.

Bug: b/410874401

+++ base/cvd/cuttlefish/host/commands/cvd/fetch/substitute.cc
@@ -114,9 +114,18 @@ Result<void> SubstituteWithMarker(const std::string& target_dir,
     }
     CF_EXPECTF(FileExists(target), "{}", target);
     // TODO: schuffelen - relax this check after migration completes
-    CF_EXPECTF(FileExists(full_link_name),
-               "Cannot substitute '{}', does not exist", full_link_name);
-    CF_EXPECTF(unlink(full_link_name.c_str()) == 0, "{}", strerror(errno));
+    const bool full_link_name_exists = FileExists(full_link_name);
+    if (symlink.ignore_link_name_presence_check()) {
+      LOG(DEBUG) << "Skipping checking for " << full_link_name
+                 << " per migration config.";
+      CF_EXPECTF(EnsureDirectoryExists(android::base::Dirname(full_link_name)),
+                 "Failed to create directory for '{}'", full_link_name);
+    } else {
+      CF_EXPECTF(FileExists(full_link_name),
+                 "Cannot substitute '{}', does not exist", full_link_name);
+      CF_EXPECTF(unlink(full_link_name.c_str()) == 0, "{}", strerror(errno));
+    }
+
     CF_EXPECT(CreateSymLink(target, full_link_name));
   }
   return {};
